% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRIENDpercentile.R
\name{FRIENDpercentile}
\alias{FRIENDpercentile}
\title{Determine FRIEND fitness percentile}
\usage{
FRIENDpercentile(VO2, age, sex, ex_mode)
}
\arguments{
\item{VO2}{The relative VO2max (ml/kg/min)}

\item{age}{Participant's age in years}

\item{sex}{Participant's sex (must begin with "m" for male and "f" for female)}

\item{ex_mode}{The exercise test mode (must begin with "t" for treadmill or "c" for cycling)}
}
\value{
Calculates age- and sex-based fitness percentiles from relative VO2max values.
}
\description{
Calculates age- and sex-based fitness percentiles from relative VO2max values.
Fitness percentiles are from the FRIEND Registry (a database of
directly-measured VO2max tests).

Treadmill fitness percentiles are from Kaminsky et al. Mayo Clin Proc 2015
Cycling fitness percentiles are from Kaminsky et al. Mayo Clin Proc 2017

\emph{When exercise test mode is unknown, the treadmill reference standard is used.}
}
