% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Cosmed_data.R
\name{read_Cosmed_data}
\alias{read_Cosmed_data}
\title{Summary of maximal Cosmed data}
\usage{
read_Cosmed_data(path)
}
\arguments{
\item{path}{Location of the Cosmed files}
}
\value{
Creates a dataframe summarizing all of the Cosmed metabolic files.
}
\description{
Creates a summary file of maximal data from a Cosmed metabolic cart.

\emph{Data from the Cosmed must be saved in .xlsx formats.}
}
\seealso{
To make it easier to define the path (find the location of the files),
use the function: \code{\link[FRIENDanalysis]{folder_location}}.
}
