% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save_data}
\alias{save_data}
\title{Save Data File}
\usage{
save_data(df, path)
}
\arguments{
\item{df}{Name of the dataframe to save.}

\item{path}{Location to save the file.}
}
\description{
Exports a data file in Excel (.xslx) format.
}
