% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRIENDpercentile.R
\docType{data}
\name{VO2}
\alias{VO2}
\title{Determine FRIEND fitness percentile}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
VO2
}
\arguments{
\item{VO2}{The relative VO2max (ml/kg/min).}

\item{age}{Participant's age (years).}

\item{sex}{Participant's sex (must begin with 'm' for male and 'f' for female - not case sensitive).}

\item{ex_mode}{The exercise test mode (must begin with 't' for treadmill or 'c' for cycling - not case sensitive).}

\item{ref_edition}{Default/blank uses standards from 2021 publication. Use '1' for original 2015 (TM) and 2017 (CY) standards.}
}
\value{
Returns age- and sex-based fitness percentile.\cr
When exercise test mode is unknown, the treadmill reference standard is used.
Percentiles above 95 are classified as 98 percent
and percentiles below 5 are classified as 3 percent.
The 2021 reference standards used are based on inclusion criteria of RER ≥ 1.10.\cr
\emph{Only returns percentiles for those aged 20-89 years
(only 20-79 years if using 2015/2017 standards).}
}
\description{
Calculates age- and sex-based fitness percentiles from relative VO2max values.\cr
Fitness percentiles are from the FRIEND Registry
(Treadmill fitness percentiles are from Kaminsky et al. Mayo Clin Proc 2015.
Cycling fitness percentiles are from Kaminsky et al. Mayo Clin Proc 2017.)
}
\keyword{datasets}
